% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Examples.R
\name{ex.FindPPIEdges}
\alias{ex.FindPPIEdges}
\title{Find PPI Edges Example}
\usage{
ex.FindPPIEdges(
  cccn_matrix,
  db_filepaths = c(),
  gm.network = NA,
  ppi.network.name
)
}
\description{
Find PPI Edges Example
}
\keyword{internal}
